requirements.txt
# MediaGuard - AI-Powered Deepfake Detection & Media Protection System
# Core dependencies

# Web Application
streamlit>=1.28.0

# Image Processing
pillow>=10.0.0
numpy>=1.24.0
imagehash>=4.3.1
opencv-python-headless>=4.8.0

# Utilities
requests>=2.32.0
beautifulsoup4>=4.12.0
pyyaml>=6.0


# Optional: For production ML models (uncomment when needed)
# tensorflow>=2.13.0
# torch>=2.0.0
# face-recognition>=1.3.0
# opencv-python>=4.8.0
scikit-learn>=1.3.0
pandas>=2.1.0
joblib>=1.3.0

azure-ai-textanalytics
azure-ai-contentsafety
azure-core



